DocuRAG/requirements.txt
streamlit
pandas
numpy
chroma
openai
requests
PyPDF2
python-docx
jsonschema
matplotlib
networkx
scikit-learn
torch
transformers
ollama
```